+++ NEWS.md
@@ -1,3 +1,8 @@
+# Ripc 0.1.0
+
+* Implement functions to pull data from the IPC API
+* Deprecate functionality to pull data direct from HDX
+
 # Ripc 0.0.1
 
 * Initial release